Assignments/01_ImageWarping/requirements.txt
gradio==4.44.1
numpy==2.1.1
opencv_python==4.8.1.78
Pillow==10.1.0

